1. Add Resample Node 2. Add a update_resample_rate API to handle the resample_rates similar to ROI

+++ rocAL/rocAL/include/loaders/audio/audio_read_and_decode.h
@@ -57,7 +57,7 @@ public:
             std::vector<uint32_t> &roi_channels,
             std::vector<uint32_t> &actual_samples,
             std::vector<uint32_t> &actual_channels,
-            std::vector<uint32_t> &actual_sample_rates);
+            std::vector<float> &actual_sample_rates);
 
     //! returns timing info or other status information
     Timing timing();
@@ -75,7 +75,7 @@ private:
     std::vector<size_t> _actual_decoded_channels;
     std::vector<size_t> _original_samples;
     std::vector<size_t> _original_channels;
-    std::vector<size_t> _original_sample_rates;
+    std::vector<float> _original_sample_rates;
     static const size_t MAX_COMPRESSED_SIZE = 1*1024*1024; // 1 Meg
     TimingDBG _file_load_time, _decode_time;
     size_t _batch_size;
